unpriv-setup: always chown roots/stacker dir
We should always chown these to the user that we're doing unpriv setup for,
otherwise they might not be able to actually use them.
